+++ backend/src/main/java/application/service/TaskService.java
@@ -21,11 +21,11 @@ public class TaskService {
         taskRepository.save(task);
     }
 
-    public Optional<Task> findById(Integer id) {
+    public Optional<Task> findById(Long id) {
         return taskRepository.findById(id);
     }
 
-    public Task setPreviousTask(Integer id, Task previousTask) {
+    public Task setPreviousTask(Long id, Task previousTask) {
         Task task = taskRepository.findById(id).get();
         task.setPreviousTask(previousTask);
